Add new buttons
